src/test/utils/metric.test.ts
import {
  metric,
  getVectorSizeSquared,
  getVectorSize,
  getTimeVelocity,
} from "../../utils/metric"

// FIXME: assign actual values
describe('metric', () => {
  it('calculates metric correctly', () => {
    const result = metric({ rs: 1, r: 2, theta: 3 });
    expect(result).toEqual([0.5, -2, -4, -36]);
  });
});

describe('getVectorSizeSquared', () => {
  it('calculates vector size squared correctly', () => {
    const vector = [1, 2, 3, 4];
    const metricData = { rs: 1, r: 2, theta: 3 };
    const result = getVectorSizeSquared(vector, metricData);
    // Replace the expected value with the actual calculation based on your functions
    expect(result).toBe(/* Your expected result */);
  });
});

describe('getVectorSize', () => {
  it('calculates vector size correctly', () => {
    const vector = [1, 2, 3, 4];
    const metricData = { rs: 1, r: 2, theta: 3 };
    const result = getVectorSize(vector, metricData);
    // Replace the expected value with the actual calculation based on your functions
    expect(result).toBe(/* Your expected result */);
  });
});

describe('getTimeVelocity', () => {
  it('calculates time velocity correctly', () => {
    const vector = [1, 2, 3];
    const metricData = { rs: 1, r: 2, theta: 3 };
    const result = getTimeVelocity(vector, metricData);
    // Replace the expected value with the actual calculation based on your functions
    expect(result).toBe(/* Your expected result */);
  });
});

